Add flag to hide credentials in output
